sensorservice: Fix init sequence for pre-1.1 API sensor HALs
The sensors API introduced a new flush() method that older binaries
won't have. For those, replace the firstFlush invocation with a
setDelay call since a lot of implementations interpreted that as
the initialization step.
We also don't want any kind of call to flush() to happen when it
isn't there, since it'll either hit a random OEM extension or
a memory address we really don't want to execute.

   sensorservice: Set the rate for pre-1.1 HALs _after_ activation

    We want to be sure these are applied in the same sequence they
    used to be...

Change-Id: I26ce923fe385751fed7d1c483a53c074249f0620

Conflicts:
	services/sensorservice/SensorService.cpp